Reimplement Charged Shot using new mana ability system
