+++ src/main/java/com/archyx/aureliumskills/mana/MAbility.java
@@ -28,7 +28,8 @@ public enum MAbility implements AbstractAbility {
     TERRAFORM(() -> Skills.EXCAVATION, 5.0, 4.0, 200, -5, 20, 20,
             new String[] {"require_sneak", "check_offhand", "sneak_offhand_bypass"}, new Object[] {false, true, true}),
     CHARGED_SHOT(() -> Skills.ARCHERY, 0.5, 0.3, 0, 0, 5, 5,
-            new String[] {"enable_message", "enable_sound"}, new Object[] {true, true}),
+            new String[] {"enable_message", "enable_sound"}, new Object[] {true, true},
+            ChargedShot.class),
     ABSORPTION(() -> Skills.DEFENSE, 2.0, 3.0, 200, -5, 10, 10,
             new String[] {"enable_particles"}, new Object[] {true},
             Absorption.class),
